Develop final pk 5041

See merge request copernico/pcs-peperina!73

+++ pcs/src/main/scala/consumers/no_registral/objeto/application/cqrs/commands/ObjetoUpdateFromObligacionHandler.scala
@@ -12,6 +12,7 @@ import consumers.no_registral.tranferencia.infrastructure.dependency_injection.O
 import cqrs.untyped.command.CommandHandler.SyncCommandHandler
 import ddd.eventCounterMax
 import design_principles.actor_model.Response
+import design_principles.actor_model.mechanism.DeliveryIdManagement.isIdempotentInternally
 
 import scala.util.{Success, Try}
 
@@ -22,6 +23,14 @@ class ObjetoUpdateFromObligacionHandler(actor: ObjetoActor, requeriment: Monitor
   ): Try[Response.SuccessProcessing] = {
     val sender = actor.context.sender()
     val obj_default: ObjetosTri = ObjetosTri(Some("None"), 0, "None", "None", "None", Some("None"), Some("None"), Some("None"), None, None, Some("None"), None, Some(0), Some("None"), Some(0), Some("None"), Some("None"), Some("None"), Some("None"), Some("None"),None,None)
+
+    log.debug(
+      f"""|CUMBIA
+          |  | command_id: ${command.deliveryId}%-20s | state_id: ${actor.state.lastDeliveryIdByEvents}%-5s
+          |  | sender    : ${actor.context.sender().path.toString.replace("akka://PersonClassificationService", "")}
+          |  | self      : ${actor.self.path.toString.replace("akka://PersonClassificationService", "")}
+          |""".stripMargin
+    )
 
     val event = ObjetoUpdatedFromObligacion(
       if (actor.state.lastDeliveryIdByEvents.equals(0)) 0 else actor.state.lastDeliveryIdByEvents,
@@ -44,23 +53,29 @@ class ObjetoUpdateFromObligacionHandler(actor: ObjetoActor, requeriment: Monitor
 
     implicit val ac: ActorSystem = actor.context.system
     val vinculoActor: ActorRef = ObjetoVinculoActor.startWithRequirements(requeriment)
+    if (isIdempotentInternally(command, actor.state.lastDeliveryIdByEvents)) {
+      log.error(
+        s"[${actor.name} | ${actor.persistenceId}] -objeto- respond idempotent because of old delivery id | $command -> " + command.deliveryId + " <= " + actor.state.lastDeliveryIdByEvents
+      )
+      sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
 
-    actor.persistEvent(event) { () =>
-      actor.state += event
-      //if (initialization != "true")
-      //  actor.informParent(command, actor.state)
-      if (actor.state.eventCounter == eventCounterMax) {
-        actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
-        actor.saveSnapshot(actor.state.copy(eventCounter = 0))
+    } else {
+      actor.persistEvent(event) { () =>
+        actor.state += event
+        //if (initialization != "true")
+        //  actor.informParent(command, actor.state)
+        if (actor.state.eventCounter == eventCounterMax) {
+          actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
+          actor.saveSnapshot(actor.state.copy(eventCounter = 0))
+        }
+
+        SendObjetoToObjetoVinculo(vinculoActor, actor, command.sujetoId, command.objetoId, command.tipoObjeto, actor.state.registro.getOrElse(obj_default).SOJ_ESTADO, requeriment, command)
+        //      actor.persistSnapshot(event, actor.state){ () =>
+        //        sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+        //
+        //      }
       }
-
-      SendObjetoToObjetoVinculo(vinculoActor, actor, command.sujetoId, command.objetoId, command.tipoObjeto, actor.state.registro.getOrElse(obj_default).SOJ_ESTADO, requeriment)
-//      actor.persistSnapshot(event, actor.state){ () =>
-//        sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
-//
-//      }
     }
-
     Success(Response.SuccessProcessing(command.aggregateRoot, command.deliveryId))
   }
 }
