+++ pcs/src/main/scala/consumers/no_registral/sujeto/application/cqrs/commands/SujetoUpdateFromObjetoHandler.scala
@@ -8,12 +8,20 @@ import consumers.no_registral.sujeto.infrastructure.dependency_injection.SujetoA
 import cqrs.untyped.command.CommandHandler.SyncCommandHandler
 import ddd.eventCounterMax
 import design_principles.actor_model.Response
+import design_principles.actor_model.mechanism.DeliveryIdManagement.isIdempotentInternally
 
 import scala.util.{Success, Try}
 
 class SujetoUpdateFromObjetoHandler(actor: SujetoActor) extends SyncCommandHandler[SujetoUpdateFromObjeto] {
   override def handle(command: SujetoUpdateFromObjeto): Try[Response.SuccessProcessing] = {
     val sender = actor.context.sender()
+    log.debug(
+      f"""|CUMBIA
+          |  | command_id: ${command.deliveryId}%-20s | state_id: ${actor.state.lastDeliveryIdByEvents}%-5s
+          |  | sender    : ${actor.context.sender().path.toString.replace("akka://PersonClassificationService", "")}
+          |  | self      : ${actor.self.path.toString.replace("akka://PersonClassificationService", "")}
+          |""".stripMargin
+    )
     val event = SujetoEvents.SujetoUpdatedFromObjeto(
       command.deliveryId,
       command.sujetoId,
@@ -24,16 +32,21 @@ class SujetoUpdateFromObjetoHandler(actor: SujetoActor) extends SyncCommandHandl
       command.clasificacionObjeto
     )
 
-    actor.persistEvent(event) { () =>
-      actor.state += event
-      SendToObjeto(actor.state, sender, actor.context, event.sujetoId, command.objetoId, command.tipoObjeto)
+    if (isIdempotentInternally(command, actor.state.lastDeliveryIdByEvents)) {
+      println(s"[${actor.name} | ${actor.persistenceId}] respond idempotent because of old delivery id | $command")
+      sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+    } else {
+      actor.persistEvent(event) { () =>
+        actor.state += event
+        SendToObjeto(actor.state, sender, actor.context, event.sujetoId, command.objetoId, command.tipoObjeto)
 
-      if (actor.state.eventCounter == eventCounterMax) {
-        actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
-        actor.saveSnapshot(actor.state.copy(eventCounter = 0))
-      }
-      actor.persistSnapshot() { _ =>
-        sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+        if (actor.state.eventCounter == eventCounterMax) {
+          actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
+          actor.saveSnapshot(actor.state.copy(eventCounter = 0))
+        }
+        actor.persistSnapshot() { _ =>
+          sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+        }
       }
     }
     Success(Response.SuccessProcessing(command.aggregateRoot, command.deliveryId))
