+++ pcs/src/main/scala/consumers/no_registral/sujeto/application/cqrs/commands/SujetoUpdateFromObjetoTreintaProcientoHandler.scala
@@ -9,15 +9,23 @@ import consumers.no_registral.sujeto.infrastructure.dependency_injection.SujetoA
 import cqrs.untyped.command.CommandHandler.SyncCommandHandler
 import ddd.eventCounterMax
 import design_principles.actor_model.Response
+import design_principles.actor_model.mechanism.DeliveryIdManagement.isIdempotentInternally
 
 import scala.util.{Success, Try}
 
 class SujetoUpdateFromObjetoTreintaProcientoHandler(actor: SujetoActor)
-  extends SyncCommandHandler[SujetoCommands.SujetoUpdateFromObjetoTreintaPorciento] {
+    extends SyncCommandHandler[SujetoCommands.SujetoUpdateFromObjetoTreintaPorciento] {
   override def handle(
-                       command: SujetoCommands.SujetoUpdateFromObjetoTreintaPorciento
-                     ): Try[Response.SuccessProcessing] = {
+      command: SujetoCommands.SujetoUpdateFromObjetoTreintaPorciento
+  ): Try[Response.SuccessProcessing] = {
     val sender: ActorRef = actor.context.sender()
+    log.debug(
+      f"""|CUMBIA
+          |  | command_id: ${command.deliveryId}%-20s | state_id: ${actor.state.lastDeliveryIdByEvents}%-5s
+          |  | sender    : ${actor.context.sender().path.toString.replace("akka://PersonClassificationService", "")}
+          |  | self      : ${actor.self.path.toString.replace("akka://PersonClassificationService", "")}
+          |""".stripMargin
+    )
     val event = SujetoEvents.SujetoUpdatedFromObjetoTreintaPorciento(
       command.deliveryId,
       command.sujetoId,
@@ -32,22 +40,25 @@ class SujetoUpdateFromObjetoTreintaProcientoHandler(actor: SujetoActor)
     //    }
 
     //    if (initialization != "true") {
-    actor.persistEvent(event) { () =>
-
-      actor.state += event
+    if (isIdempotentInternally(command, actor.state.lastDeliveryIdByEvents)) {
+      println(s"[${actor.name} | ${actor.persistenceId}] respond idempotent because of old delivery id | $command")
+      sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+    } else {
+      actor.persistEvent(event) { () =>
+        actor.state += event
         SendToObjeto(actor.state, sender, actor.context, event.sujetoId, command.objetoId, command.tipoObjeto)
 
-      if (actor.state.eventCounter == eventCounterMax) {
-        actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
-        actor.saveSnapshot(actor.state.copy(eventCounter = 0))
-      }
+        if (actor.state.eventCounter == eventCounterMax) {
+          actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
+          actor.saveSnapshot(actor.state.copy(eventCounter = 0))
+        }
 
-      actor.persistSnapshot() { _ =>
-        sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+        actor.persistSnapshot() { _ =>
+          sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
 
+        }
       }
     }
-    //    }
     Success(Response.SuccessProcessing(command.aggregateRoot, command.deliveryId))
   }
 }
